Separated sections into differenerent components
